// src/types.ts
export interface TimetableSlot {
  id: string;
  classId: string;
  stream?: string;
  gender?: string; // the specific class, e.g. 'P1', 'Senior 1'
  dayOfWeek: 'Monday' | 'Tuesday' | 'Wednesday' | 'Thursday' | 'Friday' | 'Saturday' | 'Sunday';
  startTime: string; // HH:mm
  endTime: string; // HH:mm
  subjectId: string;
  teacherId: string;
  roomId?: string; // Optional physical room
}

export interface TimetableData {
  slots: TimetableSlot[];
}